+++ src/Hooks/prepare-commit-msg.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
-php artisan hooks:prepare-commit-msg
+
+php artisan hooks:prepare-commit-msg $@
 
 STATUS_CODE=$?
 if [ ! ${STATUS_CODE} -eq 0 ]; then